+++ autosub/sub_utils.py
@@ -10,6 +10,8 @@ import json
 import gettext
 import os
 import string
+import copy
+import re
 
 # Import third-party modules
 import pysubs2
@@ -804,12 +806,77 @@ def assfile_to_txt_str(subtitles):
     return '\n'.join(event.text for event in subtitles.events)
 
 
+def split_dst_lf_src_assfile(  # pylint: disable=too-many-locals, too-many-branches
+        subtitles,
+        order=1,
+        style_name=None):
+    """
+    Split bilingual subtitles file's events automatically.
+    """
+    style_events = {}
+    event_pos = {}
+    i = 0
+    for event in subtitles.events:
+        if event.style not in style_events:
+            style_events[event.style] = [event]
+            event_pos[event.style] = i
+        else:
+            style_events[event.style].append(event)
+        i = i + 1
+
+    sorted_events_list = sorted(style_events.values(), key=len)
+    events_1 = sorted_events_list.pop()
+
+    new_ssafile = pysubs2.SSAFile()
+    new_ssafile.styles = subtitles.styles
+    new_ssafile.info = subtitles.info
+
+    new_events_1 = []
+    new_events_2 = []
+
+    if len(style_name) == 1:
+        style_name = [style_name[0], style_name[0]]
+    elif not style_name:
+        style_name = [events_1[0].style, events_1[0].style]
+
+    for event in events_1:
+        new_text_list = event.text.split(r'\N')
+        new_events_1.append(copy.deepcopy(event))
+        if len(new_text_list) == 2:
+            new_events_1[-1].text = new_text_list[0]
+            styles = re.compile(r"{\\r(.*?)}").findall(new_text_list[1])
+            new_events_1[-1].style = style_name[0]
+            new_events_2.append(copy.deepcopy(event))
+            if styles:
+                styles = styles[0].split("\\")
+                if len(styles) > 1:
+                    new_events_2[-1].text = "{\\" + new_text_list[1][4 + len(styles[0]):]
+                else:
+                    new_events_2[-1].text = new_text_list[1][4 + len(styles[0]):]
+                new_events_2[-1].style = styles[0]
+            else:
+                new_events_2[-1].text = new_text_list[1]
+                new_events_2[-1].style = style_name[1]
+
+    if order:
+        new_events = new_events_1 + new_events_2
+    else:
+        new_events = new_events_2 + new_events_1
+
+    sorted_events_list.append(new_events)
+
+    for events in sorted_events_list:
+        new_ssafile.events = new_ssafile.events + events
+
+    return new_ssafile
+
+
 def merge_bilingual_assfile(
         # pylint: disable=too-many-locals, too-many-branches, too-many-statements
         subtitles,
         order=1):
     """
-    Merge a bilingual subtitles file's events automatically.
+    Merge bilingual subtitles file's events automatically.
     """
     style_events = {}
     event_pos = {}